fix: CORUI-6264: One component per file Button (#421)
* Rename file

* Move to temp

* Restore original name

* Rename file

* Rename temp file

* Restore original

* Separate components
