+++ src/InputGroup/InputGroup.test.js
@@ -1,4 +1,4 @@
-import { Button } from '../Button/Button';
+import { Button } from '../';
 import { mount } from 'enzyme';
 import React from 'react';
 import renderer from 'react-test-renderer';
