+++ src/index.js
@@ -6,8 +6,8 @@ export { Status } from './Badge/Badge';
 export { Counter } from './Badge/Badge';
 export { Breadcrumb } from './Breadcrumb/Breadcrumb';
 export { BreadcrumbItem } from './Breadcrumb/Breadcrumb';
-export { Button } from './Button/Button';
-export { ButtonGroup } from './Button/Button';
+export { default as Button } from './Button/Button';
+export { default as ButtonGroup } from './Button/ButtonGroup';
 export { Calendar } from './Calendar/Calendar';
 export { ComboboxInput } from './ComboboxInput/ComboboxInput';
 export { DatePicker } from './DatePicker/DatePicker';
